backend/services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.2.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
